Finish Context Menu, added exception for Select
